requirements.txt
torch
torchvision
numpy
opencv-python
faiss-cpu
tqdm
streamlit
open_clip_torch